Optimized for one axes-object, nicer chart, prepared for small-multiples, release 0.2.20
The chart will now be drawn in one axes-object, so that can be used for small-multiples. Small-multiples will be better supported in the future.

Next to that, the chart looks nicer because with zorder all objects and text is better arranged. You need to compare the previous chart with the new to see the improvements.

0.2.20 will be added soon in pypi.org